requirements.txt
PySide6>=6.5.0
PySide6-Addons>=6.5.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0